Unbinded the default binding, added a new button binding, a bugfix
The bindings are now unbinded by default. A new button binding was added so that the user can choose the desired functionality when binding a button to this script. Also a bugfix was made for some cases when pauseAndStartScreensaver was used and the screensaver did not start automatically. Readme was updated.